Replace lodash for lodash.get
Don't use the full lodash module. When you do that you load the full module in memory and than import just the part you want.
When this happen you have a lot of bytes unused in memory.

With this change, the only thing in memory is the part that you use and its dependences.